added new pages and refactored the routing
