+++ views/details.hbs
@@ -7,6 +7,8 @@
       <p><span>Description:</span>{{description}}</p>
       <p><span>Difficulty level:</span>{{difficulty}}</p>
       <a class="btn" href="/attach/accessory/{{_id}}">Attach acessory</a>
+      <a class="btn" href="/edit/{{_id}}">EDIT</a>
+      <a class="btn" href="/delete/{{_id}}">Delete</a>
       <a class="btn" href="/">Back</a>
