--- views/loginPage.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Login Page</title>
-  <link href="./css/site.css" type="text/css" rel="stylesheet">
-  <link rel='shortcut icon' type='image/x-icon' href='./images/favicon.png' />
-</head>
-
-<body>
-  <div class="container">
-    <header>
-      <ul>
-        <li><img class="logo" src="./images/logo.png"></li>
-        <li><a href="/">Browse</a></li>
-        <li><a href="/">About</a></li>
-        <li><a href="/">Add a Cube</a></li>
-        <li><a href="/">Add Accessory</a></li>
-        <li><a href="/">Logout</a></li>
-        <li><a href="/">Login</a></li>
-        <li><a href="/">Register</a></li>
-      </ul>
-    </header>
-    <main>
-	<h1>Login Form</h1>
-<div class="form">
-    <form action='/' method="POST">
-        <label for="username">Username</label>
-        <input type="text" id="username" name="username">
-
-        <label for="password">Password</label>
-        <input type="password" id="password" name="password">
-        <input type="submit" value="Login">
-    </form>
-</div>
-    </main>
-  </div>
-  <footer>@Cubicle: Exercise for Express.js and Handlebars</footer>
-
-</body>
-
-</html>
