+++ views/partials/header.hbs
@@ -5,5 +5,8 @@
         <li><a href="/create">Add a Cube</a></li>
         <li><a href="/create/accessory">Create Accessory</a></li>
         <li><a href="/about">About</a></li>
+        <li><a href="/login">Log In</a></li>
+        <li><a href="/register">Register</a></li>
+        <li><a href="/logout">logout</a></li>
     </ul>
 </header>
